requirements.main.txt
openai>=1.109.1
python-dotenv>=1.1.1
requests>=2.32.5
beautifulsoup4>=4.14.2
